feat(TradeUtils): create `probabilityRowRender`
